--- src/main/java/healeat/server/web/controller/SocialLoginController.java
@@ -1,27 +1,0 @@
-package healeat.server.web.controller;
-
-import io.swagger.v3.oas.annotations.Operation;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@RestController
-@RequestMapping("/auth")
-public class SocialLoginController {
-
-    @Operation(summary = "카카오 로그인 리다이렉션", description = "카카오 로그인 경로로 리다이렉션합니다.")
-    @GetMapping("/kakao")
-    public void redirectToKakao(HttpServletResponse response) throws IOException {
-        response.sendRedirect("/oauth2/authorization/kakao");
-    }
-
-    @Operation(summary = "네이버 로그인 리다이렉션", description = "네이버 로그인 경로로 리다이렉션합니다.")
-    @GetMapping("/naver")
-    public void redirectToNaver(HttpServletResponse response) throws IOException {
-        response.sendRedirect("/oauth2/authorization/naver");
-    }
-
-}
